118 complete

+++ backend/routes/posts.js
@@ -68,9 +68,12 @@ router.put(
       content: req.body.content,
       imagePath: imagePath
     });
-    console.log(post);
-    Post.updateOne({ _id: req.params.id }, post).then((result) => {
-      res.status(200).json({ message: "Update Successful!" });
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
+      if (result.modifiedCount > 0) {
+        res.status(200).json({message: "Update successful!"});
+      } else {
+        res.status(401).json({message: "Not authorized!"});
+      };
     });
   }
 );
@@ -110,9 +113,13 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
     console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+    if (result.deletedCount > 0) {
+      res.status(200).json({message: "Deletion successful!"});
+    } else {
+      res.status(401).json({message: "Not authorized!"});
+    };
   });
 });
